Consolidate around Dark Mode name for consistency #796284012114656
