+++ app/javascript/App/components/CoinList/CoinListHeader.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import SectionHeader from '../SectionHeader'
 import Switch from '../Switch'
-import nightModeHelper from '~/nightModeHelper'
+import darkModeHelper from '~/darkModeHelper'
 
 class CoinListHeader extends Component {
   render() {
@@ -33,7 +33,7 @@ class CoinListHeader extends Component {
                 } else {
                   removeFilter('coins')
                 }
-                nightModeHelper()
+                darkModeHelper()
               }}
             />
             <div className="ml2 f5 silver">Watchlist</div>
