+++ app/javascript/App/components/Switch.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import nightModeHelper from '~/nightModeHelper'
+import darkModeHelper from '~/darkModeHelper'
 
 export default class Switch extends Component {
   state = { on: false }
@@ -12,7 +12,7 @@ export default class Switch extends Component {
     const { on } = this.state
     this.setState({ on: !on })
     onChange(!on)
-    nightModeHelper()
+    darkModeHelper()
   }
   render() {
     const { on } = this.state
