+++ app/javascript/App/darkModeHelper.js
